Drop Observable(from:record)
This observable is useful (it tracks a record's changes, completes when record is deleted, and behaves well when the a record has been unsaved yet and is thus unobservable). But it does not scale well as application complexity grows: what if you need to track several records at the same time? Conclusion: get rid of it, and restrict RxGRDB to request observation only. Request observables are just as powerful, and can already be composed efficiently.